refactor event loop handling into App
- use static methods on App rather than dealing with the global RefCell
