+++ src/gui/app.rs
@@ -1,18 +1,36 @@
 use neon::prelude::*;
-use std::time::{Duration, Instant};
 use serde_json::Value;
+use std::{
+    sync::{Arc, OnceLock},
+    iter::zip,
+    cell::RefCell,
+    thread::sleep,
+    time::{Duration, Instant},
+};
 use winit::{
     platform::pump_events::EventLoopExtPumpEvents,
     platform::run_on_demand::EventLoopExtRunOnDemand,
     event::{ElementState, KeyEvent, Event, WindowEvent},
-    event_loop::{EventLoop, ActiveEventLoop, ControlFlow},
+    event_loop::{EventLoop, EventLoopProxy, ActiveEventLoop, ControlFlow},
     keyboard::{PhysicalKey, KeyCode},
 };
 
+use crate::context::{page::Page, BoxedContext2D};
 use super::{
     event::AppEvent,
     window_mgr::WindowManager,
+    window::WindowSpec,
 };
+
+thread_local!(
+    static APP: RefCell<App> = RefCell::new(App::default());
+    static EVENT_LOOP: RefCell<EventLoop<AppEvent>> = RefCell::new(EventLoop::with_user_event().build().unwrap());
+    static PROXY: RefCell<EventLoopProxy<AppEvent>> = RefCell::new(EVENT_LOOP.with_borrow(|event_loop|
+        event_loop.create_proxy()
+    ));
+);
+
+static RENDER_CALLBACK: OnceLock<Arc<Root<JsFunction>>> = OnceLock::new();
 
 #[derive(Copy, Clone)]
 pub enum LoopMode{
@@ -35,31 +53,130 @@ impl Default for App{
     }
 }
 
-#[allow(deprecated)]
+fn add_event(event: AppEvent){
+    PROXY.with_borrow_mut(|proxy| proxy.send_event(event).ok() );
+}
+
 impl App{
-    pub fn activate<F>(&mut self, event_loop:&mut EventLoop<AppEvent>, roundtrip:F) -> bool
-        where F:FnMut(Value, Option<&mut WindowManager>) -> NeonResult<()>
-    {
-        match self.mode{
-            LoopMode::Native => {
-                let handler = self.event_handler(roundtrip);
-                event_loop.set_control_flow(ControlFlow::Wait);
-                event_loop.run_on_demand(handler).ok();
-                false
-            },
-            LoopMode::Node => {
-                let handler = self.event_handler(roundtrip);
-                event_loop.pump_events(Some(Duration::ZERO), handler);
-                self.cadence.should_continue() || !self.windows.is_empty()
+    pub fn register(callback:Root<JsFunction>){
+        RENDER_CALLBACK.get_or_init(|| Arc::new(callback));
+    }
+
+    pub fn set_mode(mode:LoopMode){
+        APP.with_borrow_mut(|app| app.mode = mode );
+    }
+
+    pub fn set_fps(fps:f32){
+        add_event(AppEvent::FrameRate(fps as u64));
+    }
+
+    pub fn open_window(spec:WindowSpec, page:Page){
+        add_event(AppEvent::Open(spec, page));
+    }
+
+    pub fn close_window(token:u32){
+        add_event(AppEvent::Close(token));
+    }
+
+    pub fn quit(){
+        APP.with_borrow_mut(|app| app.windows.remove_all() );
+        add_event(AppEvent::Quit);
+    }
+
+    #[allow(deprecated)]
+    pub fn activate(channel:Channel, deferred:neon::types::Deferred){
+        std::thread::spawn(move || {
+            loop{
+                // schedule a callback on the node event loop
+                let keep_running = channel.send(move |mut cx| {
+
+                    // define closure to relay events to js and receive canvas updates in return
+                    let roundtrip = |payload:Value, windows:Option<&mut WindowManager>| -> NeonResult<()>{
+                        let window_state = App::dispatch_events(&mut cx, payload, windows.is_some())?;
+                        if let Some(window_mgr) = windows{
+                            for (spec, page) in window_state{
+                                window_mgr.update_window(spec, page)
+                            }
+                        }
+
+                        Ok(())
+                    };
+
+                    // run the winit event loop (either once or until all windows are closed depending on mode)
+                    Ok(APP.with_borrow_mut(|app| {
+                        EVENT_LOOP.with_borrow_mut(|event_loop|{
+                            match app.mode{
+                                LoopMode::Native => {
+                                    let handler = app.event_handler(roundtrip);
+                                    event_loop.set_control_flow(ControlFlow::Wait);
+                                    event_loop.run_on_demand(handler).ok();
+                                    false
+                                },
+                                LoopMode::Node => {
+                                    let handler = app.event_handler(roundtrip);
+                                    event_loop.pump_events(Some(Duration::ZERO), handler);
+                                    app.cadence.should_continue() || !app.windows.is_empty()
+                                }
+                            }
+                        })
+                    }))
+                }).join();
+
+                // in node-events mode, wait briefly before checking for new events
+                match keep_running{
+                    Ok(true) => sleep(Duration::from_millis(1)),
+                    _ => break
+                }
+            }
+
+            // resolve the promise
+            deferred.settle_with(&channel, move |mut cx| Ok(cx.undefined()) );
+        });
+    }
+
+    fn dispatch_events(cx:&mut TaskContext, payload:Value, is_render:bool) -> NeonResult<Vec<(WindowSpec, Page)>>{
+        // send payload to js for event dispatch and canvas drawing
+        let mut call = match RENDER_CALLBACK.get(){
+            None => return Ok(vec![]),
+            Some(callback)=> callback.to_inner(cx).call_with(cx),
+        };
+        call.arg(cx.boolean(is_render))
+            .arg(cx.string(payload.to_string()));
+
+        match is_render{
+            true => {
+                // for a full roundtrip, pass events to js then unpack updated window specs & contexts
+                let response = call.apply::<JsValue, _>(cx)?
+                    .downcast::<JsArray, _>(cx).or_throw(cx)?
+                    .to_vec(cx)?;
+
+                let specs_json = response[0].downcast::<JsString, _>(cx).or_throw(cx)?.value(cx);
+                let specs:Vec<WindowSpec> = serde_json::from_str(&specs_json)
+                    .or_else(|err| cx.throw_error(format!("Malformed response from window event handler: {}", err)) )?;
+
+                let contexts = response[1].downcast::<JsArray, _>(cx).or_throw(cx)?.to_vec(cx)?;
+                let pages = contexts.iter().map(|boxed|
+                    boxed.downcast::<BoxedContext2D, _>(cx).ok()
+                        .map(|ctx| ctx.borrow().get_page())
+                );
+
+                // group spec + page pairs for each window
+                let window_state = zip(specs, pages)
+                    .filter_map(|(spec, page)| page.map(|page| (spec, page) ))
+                    .collect();
+
+                Ok(window_state)
+            }
+
+            false => {
+                // if this is just a UI-event delivery pass, ignore the js callback's return value
+                call.exec(cx)?;
+                Ok(vec![])
             }
         }
     }
 
-    pub fn close_all(&mut self){
-        self.windows.remove_all();
-    }
-
-    pub fn event_handler<F>(&mut self, mut roundtrip:F) -> impl FnMut(Event<AppEvent>, &ActiveEventLoop) + use<'_, F>
+    fn event_handler<F>(&mut self, mut roundtrip:F) -> impl FnMut(Event<AppEvent>, &ActiveEventLoop) + use<'_, F>
         where F:FnMut(Value, Option<&mut WindowManager>) -> NeonResult<()>
     {
         move |event, event_loop| match event {
